Add coordinator to show passcode
