+++ Trust/Lock/Coordinators/LockEnterPasscodeCoordinator.swift
@@ -1,0 +1,23 @@
+// Copyright SIX DAY LLC. All rights reserved.
+
+import UIKit
+
+class LockEnterPasscodeCoordinator: Coordinator {
+    var coordinators: [Coordinator] = []
+    private let window: UIWindow
+    private let model: LockEnterPasscodeViewModel
+    private lazy var lockEnterPasscodeViewController: LockEnterPasscodeViewController = {
+        return LockEnterPasscodeViewController(model: model)
+    }()
+    init(window: UIWindow, model: LockEnterPasscodeViewModel) {
+        self.window = window
+        self.model = model
+    }
+    func start() {
+        window.rootViewController = lockEnterPasscodeViewController
+        window.makeKeyAndVisible()
+    }
+    func dismiss() {
+        window.isHidden = true
+    }
+}
